requirements.txt
council-ai==0.0.7